+++ Cargo.toml
@@ -11,14 +11,11 @@ license = "MIT"
 exclude = [".*"]
 
 [features]
-# FIXME: docs in `ser` mod depend on geozero
-# default = ["geozero"]
 shp = ["shapefile"]
-fgb = ["flatgeobuf"] # , "geozero"
+fgb = ["flatgeobuf"]
 
 [dependencies]
 geo-types = { version = "0.7.11" }
-# geozero = { version = "0.14.0", optional = true }
 serde = { version = "1.0.188" }
 geoserde_derive = { version = "0.1.0", path = "derive" }
 geojson = { version = "0.24.2", optional = true }
