+++ tests/mod.rs
@@ -1,1 +1,2 @@
-mod v2;
+mod v0_5;
+mod v0_6;
