+++ tests/v0_5/geometry_detection.rs
