+++ tests/v0_5/mod.rs
@@ -1,0 +1,15 @@
+mod feature_serializer;
+mod fgb;
+mod geometry_detection;
+#[cfg(feature = "geozero")]
+mod line;
+#[cfg(feature = "geozero")]
+mod line_string;
+#[cfg(feature = "geozero")]
+mod point;
+#[cfg(feature = "geozero")]
+mod polygon;
+#[cfg(feature = "geozero")]
+mod property_array;
+#[cfg(feature = "geozero")]
+mod rect;
