+++ tests/v0_5/property_array.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "geozero")]
-
 use geoserde::PropertySerializer;
 use serde::Serialize;
