+++ tests/v0_6/duplicate_field.rs
