+++ tests/v0_6/shp.rs
@@ -1,7 +1,8 @@
 #[test]
 fn shape_test() {
     let mut reader =
-        shapefile::Reader::from_path("/Users/p4/Documents/geoserde/tests/v2/sample/a.shp").unwrap();
+        shapefile::Reader::from_path("/Users/p4/Documents/geoserde/tests/v0_6/sample/a.shp")
+            .unwrap();
     // ジオメトリは1コピーとなる
     for res in reader.iter_shapes_and_records_as::<shapefile::Polyline, MyStruct>() {
         let (geom, prop) = res.unwrap();
